# chore(parser): optimize the most common lexer matches into lookup tables