## Overview
Basically just bastardizes the good ideas in https://github.com/maciejhirsz/logos in a modest change trying to get the most value without making the code unreadable or hard to maintain. Both functions internally call `is_ascii` to prevent overflows since we're dealing with utf-8.

There's obviously more you could do but from my profiles this seems to be the most impactful paths without making the whole thing unreadable.

Feel free to close as there are obvious tradeoffs here but I opened for your consideration.

## Impact
The impact of this change really depends on the document being parsed. For small documents its negligible but for large documents, particularly schema SDLs the change speeds things up to ~18%.

```
test result: ok. 0 passed; 0 failed; 65 ignored; 0 measured; 0 filtered out; finished in 0.00s

     Running benches/query.rs (/Users/allancalix/src/github.com/allancalix/apollo-rs/target/release/deps/query-008ff12600681735)
Benchmarking many_aliases: Warming up for 3.0000 s
Warning: Unable to complete 100 samples in 5.0s. You may wish to increase target time to 9.8s, enable flat sampling, or reduce sample count to 50.
many_aliases            time:   [1.9430 ms 1.9473 ms 1.9521 ms]
                        change: [-2.1904% -1.8069% -1.4314%] (p = 0.00 < 0.05)
                        Performance has improved.
Found 6 outliers among 100 measurements (6.00%)
  4 (4.00%) high mild
  2 (2.00%) high severe

query_lexer             time:   [1.1066 µs 1.1090 µs 1.1115 µs]
                        change: [-6.5579% -6.1115% -5.7380%] (p = 0.00 < 0.05)
                        Performance has improved.
Found 4 outliers among 100 measurements (4.00%)
  1 (1.00%) low mild
  1 (1.00%) high mild
  2 (2.00%) high severe

query_parser            time:   [7.4054 µs 7.4149 µs 7.4241 µs]
                        change: [-13.829% -10.061% -6.8751%] (p = 0.00 < 0.05)
                        Performance has improved.
Found 2 outliers among 100 measurements (2.00%)
  2 (2.00%) high mild

     Running benches/supergraph.rs (/Users/allancalix/src/github.com/allancalix/apollo-rs/target/release/deps/supergraph-4c8cf367fca5ad96)
supergraph_lexer        time:   [47.780 µs 47.864 µs 47.956 µs]
                        change: [-5.9375% -5.6993% -5.4348%] (p = 0.00 < 0.05)
                        Performance has improved.
Found 4 outliers among 100 measurements (4.00%)
  1 (1.00%) low mild
  2 (2.00%) high mild
  1 (1.00%) high severe

supergraph_parser       time:   [201.49 µs 201.96 µs 202.47 µs]
                        change: [-1.8449% -1.4664% -1.0951%] (p = 0.00 < 0.05)
                        Performance has improved.
Found 4 outliers among 100 measurements (4.00%)
  2 (2.00%) high mild
  2 (2.00%) high severe
```

The delta on the public github schema was quite sizable:
```
apollo - parse github schema
                        time:   [4.8513 ms 4.8560 ms 4.8611 ms]
                        change: [-18.862% -18.703% -18.548%] (p = 0.00 < 0.05)
                        Performance has improved.
Found 2 outliers among 100 measurements (2.00%)
  1 (1.00%) high mild
  1 (1.00%) high severe
```